refactor(protocol-designer): move pipettes into own dir
