+++ protocol-designer/src/file-data/actions.js
@@ -1,6 +1,5 @@
 // @flow
 import type {FileMetadataFieldAccessors} from './types'
-import type {PipetteName} from './pipetteData'
 
 export const updateFileMetadataFields = (payload: {[accessor: FileMetadataFieldAccessors]: string}) => ({
   type: 'UPDATE_FILE_METADATA_FIELDS',
@@ -9,10 +8,5 @@ export const updateFileMetadataFields = (payload: {[accessor: FileMetadataFieldA
 
 export const saveFileMetadata = (payload: {[accessor: FileMetadataFieldAccessors]: string}) => ({
   type: 'SAVE_FILE_METADATA',
-  payload
-})
-
-export const updatePipettes = (payload: {['left' | 'right']: ?PipetteName}) => ({
-  type: 'UPDATE_PIPETTES',
   payload
 })
