+++ protocol-designer/src/file-data/selectors/index.js
@@ -2,4 +2,3 @@
 export * from './commands'
 export * from './fileCreator'
 export * from './fileFields'
-export * from './pipettes'
