+++ protocol-designer/src/pipettes/actions.js
@@ -1,0 +1,7 @@
+// @flow
+import type {PipetteName} from './pipetteData'
+
+export const updatePipettes = (payload: {['left' | 'right']: ?PipetteName}) => ({
+  type: 'UPDATE_PIPETTES',
+  payload
+})
