+++ protocol-designer/src/pipettes/index.js
@@ -1,7 +1,6 @@
 // @flow
-/** This is the big selector that generates a .json file to download */
 import * as actions from './actions'
-import {rootReducer, type RootState} from './reducers'
+import {rootReducer} from './reducers'
 import * as selectors from './selectors'
 export * from './types'
 
@@ -9,8 +8,4 @@ export {
   actions,
   rootReducer,
   selectors
-}
-
-export type {
-  RootState
 }
