+++ protocol-designer/src/top-selectors/substeps.js
@@ -2,7 +2,7 @@
 import {createSelector} from 'reselect'
 import mapValues from 'lodash/mapValues'
 
-import {equippedPipettes} from '../file-data/selectors/pipettes'
+import {selectors as pipetteSelectors} from '../pipettes'
 import {selectors as labwareIngredSelectors} from '../labware-ingred/reducers'
 import {selectors as steplistSelectors} from '../steplist'
 import {selectors as fileDataSelectors} from '../file-data'
@@ -19,7 +19,7 @@ import type {StepSubItemData} from '../steplist/types'
 
 export const allSubsteps: Selector<{[StepIdType]: StepSubItemData | null}> = createSelector(
   steplistSelectors.validatedForms,
-  equippedPipettes,
+  pipetteSelectors.equippedPipettes,
   labwareIngredSelectors.getLabware,
   namedIngredsByLabware,
   steplistSelectors.orderedSteps,
