+++ protocol-designer/src/well-selection/actions.js
@@ -4,7 +4,7 @@ import selectors from './selectors'
 import {changeFormInput} from '../steplist/actions'
 
 import {selectors as steplistSelectors} from '../steplist/reducers'
-import {selectors as fileDataSelectors} from '../file-data'
+import {selectors as pipetteSelectors} from '../pipettes'
 import {selectors as labwareIngredSelectors} from '../labware-ingred/reducers'
 
 import type {ThunkDispatch, GetState} from '../types'
@@ -62,7 +62,7 @@ export const openWellSelectionModal = (payload: OpenWellSelectionModalPayload) =
     // initially selected wells in form get selected in state before modal opens
     dispatch(selectWells(wells))
 
-    const pipettes = fileDataSelectors.equippedPipettes(state)
+    const pipettes = pipetteSelectors.equippedPipettes(state)
     const labware = labwareIngredSelectors.getLabware(state)
     // TODO type this action, make an underline fn action creator
